v5.1.5 patch --- - Bugs fixed:   - Prevent running code in EARD report plug-in misc method if the event type is not WF_APPLICATION   - Prevent closing a popen structure if the process is already death.   - file_ module's function naming changed to ear_file to prevent symbol name collisions with third-party applications.

+++ src/common/system/popen.c
@@ -130,17 +130,19 @@ state_t popen_open(char *command, int escape_lines, int one_shot, popen_t *p)
 
 void popen_close(popen_t *p)
 {
-	debug("killing %d", p->pid);
+	if (p && p->opened) {
+		debug("killing %d", p->pid);
 
-	// SIGKILL might be too aggressive but it can not be blocked.
-	kill(p->pid, SIGKILL);
+		// SIGKILL might be too aggressive but it can not be blocked.
+		kill(p->pid, SIGKILL);
 
-	fclose(p->file);
+		fclose(p->file);
 
-	waitpid(p->pid, NULL, 0);
+		waitpid(p->pid, NULL, 0);
 
-	// p->opened = 0;
-	memset(p, 0, sizeof(popen_t));
+		// p->opened = 0;
+		memset(p, 0, sizeof(popen_t));
+	}
 }
 
 static void static_flush(popen_t *p, char *buffer)
